sensor_publisher/requirements.txt
RPI.BME280
paho-mqtt
RPI.GPIO